Changed UserId to long instead of guid. Updated Database migrations appropriately

+++ ChatApplication.Data.Contracts/Repositories/Repository.cs
@@ -17,7 +17,7 @@ namespace ChatApplication.Data.Contracts.Repositories
             ReaderDelegate = reader;
             WriterDelegate = writer;
         }
-        public TEntity Get(int id)
+        public TEntity Get(long id)
         {
             return ReaderDelegate.Get(id);
         }
